+++ IMU/src/main/java/org/nowireless/imu/ImmutableIMUData.java
@@ -5,10 +5,10 @@ public class ImmutableIMUData implements IMUData {
 	private final double y;
 	private final double z;
 	private final double angle;
-	private final int timestamp;
+	private final long timestamp;
 	
 	
-	public ImmutableIMUData(double x, double y, double z, double angle, int ts) {
+	public ImmutableIMUData(double x, double y, double z, double angle, long ts) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
@@ -37,7 +37,7 @@ public class ImmutableIMUData implements IMUData {
 	}
 
 	@Override
-	public int getTimeStamp() {
+	public long getTimeStamp() {
 		return timestamp;
 	}
 }
